driver: use C++11 mutex and guard lock.
Instead of the Boost version of scoped_lock.

Should work the same way (only a single mutex to lock in each scope) and reduces Boost usage.

The enable/disable services can be invoked concurrently with other usage of the connection, and SimpleMessage is not thread safe, so enforce serialisation here.